refactor(onboard): extract summary formatting
